+++ cpp_version/EVAP.h
@@ -84,8 +84,8 @@ class EVAP {
         // Constructors
         //EVAP(double lat, int n, double elv);  // default constructor
         //EVAP(double lat, int n, double elv, int y);  // default w/ year
-        EVAP(double lat, int n, double elv, int y, float sf, float tc, 
-             float sw);
+        EVAP(double lat, int n, double elv, int y, double sf, double tc, 
+             double sw);
         
         // Get Variable Functions:
         double get_ho();
